refactor(17451): refactor default option - add sorting function
